requirements.txt
astroid==3.3.8
autopep8==2.3.2
colorama==0.4.6
coverage==7.6.10
dill==0.3.9
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pylint==3.3.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytest==8.3.4
tomlkit==0.13.2
